feat(signup): adds ui and functionality for signup
This commit adds functionality for signing up a user. It includes a
UI, a controller function and a simple dashboard where a user can
be directed to the sign up page after which they will be directed
to the homepage if sign up is successful or back to the form if sign
up fails.

[Finishes #158656372]