feat(sign-in): adds sign in functionality
This commit adds sign in functionality. It adds a UI where
a user can access the sign in page and fill in their email
and password. This page displays error if the email is
not registered or the password is wrong.

Once a user registers successfully, they are redirected to
the application homepage.

[Finishes #158657591]