feat(logout): adds logout feature
This commit adds the ability for a user to logout.
It modifies the navigation bar to contain drop
down that has this option.

It also moves methods for flash and validation to
a helper file.

It changes up the dashboard to give it better styling.

It also adds auth guarding to prevent access to dash-
board for logged in users and access to homepage for
non-logged in users.

[Finishes ]