feat(signin-validation): adds validation errors
This commit adds validation errors for sign in
functionality. It shows errors if the email is
invalid or blank and if password is less than 6
characters.

It also adds a couple of helper methods for to
help only accept accepted parameters.

It also adds tests for the sign in validation
errors.

[Finishes #158657591]