feat(topics): gives user ability to select topics
This commit gives the user the ability to select
topics based on various categories

It adds the category controller to the application
which gives the user options to select after which
they are redirected to the topics page for the
category.

It also adds a many to many relationship between
users and topics and one to many relationship between
categories and topics.

[Finishes #158656883]