feat(homepage): add sent articles to homepage
This commit adds a view of sent articles to the homepage. It gives the
user all the articles that have been sent as history.

It also adds a close and add button beside the users topics on the
homepage which enables a user to add or remove preferred topics more
easily.

[Finishes #158657278]