feat(homepage): adds history table to homepage
This commit adds a table that displays the user's history
in regards to the articles that have been emailed to the
user.

It also adds the google custom search api to server users
with articles and urls.

It also adds the ability for a user to delete a topic they
are interested in by adding a remove button on the topics.

[Finishes #158657278]